Merge Fixes
I almost thought that the entire project is doomed after my merge didn't go so well. I chose to rebase and it didn't go so well.

But I fixed it in the end by copying folder from my test branch to another location and then copying all files and folders inside that folder and pasting them into the folder with repository and it seems like it worked.

You might notice that I'm committing this change again. The reason for doing so is that the previous commit was made prematurely.